PageInfoText above the search result, Paginations below, dark mode id color
