+++ src/app/components/Pokemon/PokemonNameSection.tsx
@@ -11,7 +11,7 @@ const PokemonNameSection = ({ name, id }: Props) => {
   return (
     <>
       <h2 className="sm:pt-6">{capitalizedName}</h2>
-      <p className="text-gray-500">#{idToDisplay}</p>
+      <p className="text-gray-500 dark:text-gray-200">#{idToDisplay}</p>
     </>
   );
 };
