+++ src/app/components/ui/Paginator.tsx
@@ -9,8 +9,8 @@ const Paginator = ({ onClick, className, disabled, buttonText }: PaginatorProps)
   return (
     <button 
       onClick={onClick} 
-      className={`px-4 py-2 bg-[var(--check-blue)] text-white rounded ${
-        disabled ? 'opacity-50' : 'cursor-pointer hover:scale-105 hover:shadow-lg font-bold'
+      className={`px-3 py-2 bg-[var(--check-blue)] text-white rounded ${
+        disabled ? 'opacity-50' : 'cursor-pointer hover:scale-105 hover:font-semibold hover:shadow-lg text-xl'
       } ${className}`}
       type="button"
       disabled={disabled}
